requirements.txt
streamlit
pandas
sklearn.preprocessing
pickle
